Arquivo handlers/users/http.go + model.go  conciliados manualmente como handlers/http/users.go
