+++ code/internal/handlers/http/users.go
@@ -1,19 +1,18 @@
-package users
+package http
 
 import (
 	"context"
 	"net/http"
 
+	"github.com/SOAT1StackGoLang/tech-challenge/helpers"
+	"github.com/SOAT1StackGoLang/tech-challenge/internal/core/domain"
 	"github.com/SOAT1StackGoLang/tech-challenge/internal/core/ports"
 	restfulspec "github.com/emicklei/go-restful-openapi/v2"
 	restful "github.com/emicklei/go-restful/v3"
 )
 
-var (
-	ctx context.Context
-)
-
 type UserHandler struct {
+	ctx          context.Context
 	usersUseCase ports.UsersUseCase
 }
 
@@ -27,12 +26,42 @@ type IDUser struct {
 	ID string `xml:"id" json:"id" description:"ID do cliente no sistema"`
 }
 
+type User struct {
+	ID       string `xml:"id" json:"id" description:"ID do cliente no sistema"`
+	Document string `xml:"document" json:"document" description:"CPF do cliente"`
+	Name     string `xml:"name" json:"name" description:"Nome do cliente"`
+	Email    string `xml:"email" json:"email" description:"Email do cliente"`
+}
+
+func (u *User) fromDomain(user *domain.User) {
+	if u == nil {
+		u = &User{}
+	}
+
+	u.ID = user.ID.String()
+	u.Name = user.Name
+	u.Document = user.Document
+	u.Email = user.Email
+}
+
+func (u *User) toDomain() *domain.User {
+	if u == nil {
+		u = &User{}
+	}
+
+	return &domain.User{
+		ID:       helpers.SafeUUIDFromString(u.ID),
+		Document: u.Document,
+		Name:     u.Name,
+		Email:    u.Email,
+	}
+}
+
 func NewUserHandler(
-	mainCtx context.Context,
+	ctx context.Context,
 	useCase ports.UsersUseCase,
 	ws *restful.WebService,
 ) *UserHandler {
-	ctx = mainCtx
 	handler := &UserHandler{
 		usersUseCase: useCase,
 	}
@@ -68,7 +97,7 @@ func (uH *UserHandler) Create(req *restful.Request, resp *restful.Response) {
 		return
 	}
 
-	result, err := uH.usersUseCase.CreateUser(ctx, user.Name, user.Document, user.Email)
+	result, err := uH.usersUseCase.CreateUser(uH.ctx, user.Name, user.Document, user.Email)
 	if err != nil {
 		_ = resp.WriteError(http.StatusInternalServerError, err)
 		return
@@ -82,7 +111,7 @@ func (uH *UserHandler) Create(req *restful.Request, resp *restful.Response) {
 func (uH *UserHandler) Validate(req *restful.Request, resp *restful.Response) {
 	document := req.PathParameter("document")
 
-	result, err := uH.usersUseCase.ValidateUser(ctx, document)
+	result, err := uH.usersUseCase.ValidateUser(uH.ctx, document)
 	if err != nil {
 		_ = resp.WriteError(http.StatusInternalServerError, err)
 		return
